Add shocks to SFC models

The sfcr_scenario function is used to add a shock to a SFC model generated with the sfcr_sim function.

sfcr_scenario(
  steady_state,
  equations,
  t = 100,
  exogenous,
  parameters,
  shock_exg = NULL,
  shock_param = NULL
)

Arguments

steady_state
A simulated SFC model with sfcr_sim function to provide the steady state values of the variables.

equations
A list containing all the equations of the model to be simulated. See details.

t
A number specifying the total number of periods of the model to be simulated. It should be at least 2 periods.

exogenous, parameters
Named lists with the initial values of the exogenous and parameters. This should be exactly the same values as supplied to sfcr_sim.

shock_exg, shock_param
Named lists with the new values for the exogenous and/or parameters that are being shocked to calculate a different scenario.

Details

The output of sfcr_sim will contain a tibble with the exogenous and endogenous variables, as well as the parameters, of the simulated model as columns.

The equations at the equations argument should take the usual R formula syntax. Also, endogenous and exogenous variables should explicitly indicate if they're to be considered contemporaneously or with lags, while parameters in should not be accompanied by period of evaluation indication.

On the other hand, the named lists should not include time indices.

See sfcr_sim for further information.

Examples

eqs <- list(
  TX_s[t] ~ TX_d[t],
  YD[t] ~ W[t] * N_s[t] - TX_s[t],
  C_d[t] ~ alpha1 * YD[t] + alpha2 * H_h[t-1],
  H_h[t] ~ YD[t] - C_d[t] + H_h[t-1],
  N_s[t] ~ N_d[t],
  N_d[t] ~ Y[t] / W[t],
  C_s[t] ~ C_d[t],
  G_s[t] ~ G_d[t],
  Y[t] ~ C_s[t] + G_s[t],
  TX_d[t] ~ theta * W[t] * N_s[t],
  H_s[t] ~ G_d[t] - TX_d[t] + H_s[t-1]
)

exg <- list("G_d" = 20, "W" = 1)
params <- list("alpha1" = 0.6, "alpha2" = 0.4, "theta" = 0.2)

# t is set to 10 to run faster. A usual model should run at least 50 periods to find a steady state
steady_state <- sfcr_sim(equations = eqs, t = 10, exogenous = exg,
                         parameters = params, initial = NULL)

# Increase G_d from 20 to 30
sfcr_scenario(steady_state, eqs, t = 10, exg, params, shock_exg = list("G_d" = 30))

#> # A tibble: 10 x 17
#>        t  TX_s    YD   C_d   H_h   N_s   N_d   C_s   G_s     Y  TX_d   H_s   G_d
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1  16.8  67.2  64.0  62.4  84.0  84.0  64.0    20  84.0  16.8  62.4    20
#>  2     2  17.3  69.2  66.5  65.1  86.5  86.5  66.5    20  86.5  17.3  65.2    20
#>  3     3  17.7  70.8  68.6  67.4  88.6  88.6  68.6    20  88.6  17.7  67.4    20
#>  4     4  18.1  72.3  70.3  69.4  90.3  90.3  70.3    20  90.3  18.1  69.4    20
#>  5     5  18.4  73.4  71.8  71.0  91.8  91.8  71.8    20  91.8  18.4  71.0    20
#>  6     6  22.5  89.8  82.3  78.5 112.  112.   82.3    30 112.   22.5  78.6    30
#>  7     7  23.6  94.5  88.1  84.9 118.  118.   88.1    30 118.   23.6  84.9    30
#>  8     8  24.6  98.4  93.0  90.3 123.  123.   93.0    30 123.   24.6  90.3    30
#>  9     9  25.4 102.   97.2  94.9 127.  127.   97.2    30 127.   25.4  94.9    30
#> 10    10  26.1 105.  101.   98.7 131.  131.  101.     30 131.   26.1  98.8    30
#> # … with 4 more variables: W <dbl>, alpha1 <dbl>, alpha2 <dbl>, theta <dbl>

# Increase alpha1 from 0.6 to 0.8
sfcr_scenario(steady_state, eqs, t = 10, exg, params, shock_param = list("alpha1" = 0.8))

#> # A tibble: 10 x 17
#>        t  TX_s    YD   C_d   H_h   N_s   N_d   C_s   G_s     Y  TX_d   H_s   G_d
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1  16.8  67.2  64.0  62.4  84.0  84.0  64.0    20  84.0  16.8  62.4    20
#>  2     2  17.3  69.2  66.5  65.1  86.5  86.5  66.5    20  86.5  17.3  65.2    20
#>  3     3  17.7  70.8  68.6  67.4  88.6  88.6  68.6    20  88.6  17.7  67.4    20
#>  4     4  18.1  72.3  70.3  69.4  90.3  90.3  70.3    20  90.3  18.1  69.4    20
#>  5     5  18.4  73.4  71.8  71.0  91.8  91.8  71.8    20  91.8  18.4  71.0    20
#>  6     6  26.9 107.  114.   64.1 134.  134.  114.     20 134.   26.9  64.2    20
#>  7     7  25.3 101.  107.   58.7 127.  127.  107.     20 127.   25.3  58.8    20
#>  8     8  24.1  96.5 101.   54.5 121.  121.  101.     20 121.   24.1  54.7    20
#>  9     9  23.2  92.8  96.1  51.3 116.  116.   96.1    20 116.   23.2  51.5    20
#> 10    10  22.5  90.0  92.5  48.8 112.  112.   92.5    20 112.   22.5  49.0    20
#> # … with 4 more variables: W <dbl>, alpha1 <dbl>, alpha2 <dbl>, theta <dbl>

# Increase W to 2, alpha2 to 0.5, and decrease theta to 0.15
sfcr_scenario(steady_state, eqs, t = 10, exg, params, shock_exg = list("W" = 2),
              shock_param = list("alpha2" = 0.5, "theta" = 0.15))

#> # A tibble: 10 x 17
#>        t  TX_s    YD   C_d   H_h   N_s   N_d   C_s   G_s     Y  TX_d   H_s   G_d
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1  16.8  67.2  64.0  62.4  84.0  84.0  64.0    20  84.0  16.8  62.4    20
#>  2     2  17.3  69.2  66.5  65.1  86.5  86.5  66.5    20  86.5  17.3  65.2    20
#>  3     3  17.7  70.8  68.6  67.4  88.6  88.6  68.6    20  88.6  17.7  67.4    20
#>  4     4  18.1  72.3  70.3  69.4  90.3  90.3  70.3    20  90.3  18.1  69.4    20
#>  5     5  18.4  73.4  71.8  71.0  91.8  91.8  71.8    20  91.8  18.4  71.0    20
#>  6     6  17.0  96.3  93.3  74.0  56.6  56.6  93.3    20 113.   17.0  74.0    20
#>  7     7  17.5  98.9  96.3  76.6  58.2  58.2  96.3    20 116.   17.5  76.6    20
#>  8     8  17.8 101.   98.9  78.7  59.5  59.5  98.9    20 119.   17.8  78.7    20
#>  9     9  18.2 103.  101.   80.6  60.6  60.6 101.     20 121.   18.2  80.6    20
#> 10    10  18.5 105.  103.   82.1  61.5  61.5 103.     20 123.   18.5  82.1    20
#> # … with 4 more variables: W <dbl>, alpha2 <dbl>, theta <dbl>, alpha1 <dbl>