Simulate a stock-flow consistent model

The sfcr_sim function is used to simulate a SFC model. With adequate number of periods, the simulated model should converge to a steady state scenario.

sfcr_sim(
  equations,
  t = 100,
  exogenous,
  parameters,
  initial = NULL,
  hidden = NULL
)

Arguments

equations
A list containing all the equations of the model to be simulated. See details.

t
A number specifying the total number of periods of the model to be simulated. It should be at least 2 periods.

exogenous, parameters, initial
Named lists with exogenous, parameters, and initial values of endogenous variables as values. See details.

hidden
Named list that identify the two variables that make the hidden equality in the SFC model, e.g., list("H_h" = "H_s"). Defaults to NULL. If hidden is supplied, the model will evaluate if the hidden equation is satisfied. If it is not, it will throw out an error.

Details

The output of sfcr_sim() will contain a tibble with the exogenous and endogenous variables, as well as the parameters, of the simulated model as columns.

The equations at the equations argument should take the usual R formula syntax. Also, endogenous and exogenous variables should explicitly indicate if they're to be considered contemporaneously or with lags, while parameters in should not be accompanied by period of evaluation indication. For example, a consumption equation should be written like C_d[t] ~ alpha1 * YD[t] + alpha2 * H_h[t-1].

The named lists should not include time indications. See examples.

Examples

eqs <- list(
  TX_s[t] ~ TX_d[t],
  YD[t] ~ W[t] * N_s[t] - TX_s[t],
  C_d[t] ~ alpha1 * YD[t] + alpha2 * H_h[t-1],
  H_h[t] ~ YD[t] - C_d[t] + H_h[t-1],
  N_s[t] ~ N_d[t],
  N_d[t] ~ Y[t] / W[t],
  C_s[t] ~ C_d[t],
  G_s[t] ~ G_d[t],
  Y[t] ~ C_s[t] + G_s[t],
  TX_d[t] ~ theta * W[t] * N_s[t],
  H_s[t] ~ G_d[t] - TX_d[t] + H_s[t-1]
)

exg <- list("G_d" = 20, "W" = 1)
params <- list("alpha1" = 0.6, "alpha2" = 0.4, "theta" = 0.2)

# t is set to 10 to run faster. A usual model should run at least 50 periods to find a steady state
sfcr_sim(equations = eqs, t = 10, exogenous = exg, parameters = params, initial = NULL)

#> # A tibble: 10 x 17
#>        t  TX_s    YD   C_d   H_h   N_s   N_d   C_s   G_s     Y  TX_d   H_s   G_d
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1     1  1      1     1     1     1     1     1       1   1    1      1       1
#>  2     2  7.85  31.4  19.2  13.2  39.2  39.2  19.2    20  39.2  7.85  13.2    20
#>  3     3  9.72  38.9  28.6  23.4  48.6  48.6  28.6    20  48.6  9.72  23.4    20
#>  4     4 11.3   45.2  36.5  32.1  56.5  56.5  36.5    20  56.5 11.3   32.1    20
#>  5     5 12.6   50.5  43.2  39.5  63.2  63.2  43.2    20  63.2 12.6   39.5    20
#>  6     6 13.8   55.1  48.8  45.7  68.8  68.8  48.8    20  68.8 13.8   45.7    20
#>  7     7 14.7   58.9  53.6  51.0  73.6  73.6  53.6    20  73.6 14.7   51.0    20
#>  8     8 15.5   62.1  57.7  55.5  77.7  77.7  57.7    20  77.7 15.5   55.5    20
#>  9     9 16.2   64.9  61.1  59.2  81.1  81.1  61.1    20  81.1 16.2   59.3    20
#> 10    10 16.8   67.2  64.0  62.4  84.0  84.0  64.0    20  84.0 16.8   62.4    20
#> # … with 4 more variables: W <dbl>, alpha1 <dbl>, alpha2 <dbl>, theta <dbl>